refactor(wallet): rename Controller to Wallet and move to wallet package
